+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/).
 
 ## [Unreleased]
+### Added
+- Use workalendar to avoid notifications on holidays
 ### Changed
 - Better webdriver checking and error handling
 ### Fixed